pyproject.toml
[tool.poetry]
name = "webservice"
version = "0.1.0"
description = ""
authors = ["alien <t108368530@ntut.org.tw>"]

[tool.poetry.dependencies]
python = "^3.8"
tensorflow-gpu = "2.3.1"
librosa = "^0.8.0"
fastapi = "^0.61.1"
ujson = "^4.0.1"
uvicorn = "^0.12.1"
gunicorn = "^20.0.4"
PyYAML = "^5.3.1"
pydantic = {extras = ["dotenv"], version = "^1.6.1"}
loguru = "^0.5.3"
uvloop = "^0.14.0"
httptools = "^0.1.1"
python-multipart = "^0.0.5"
redis = "^3.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
requests = "^2.24.0"
black = {version = "^20.8b1", allow-prereleases = true}
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
